Refactor MainPage and update navigation in MainWindow
Changed MainPage namespace and layout to MainForPage.
Restored using directives in MainPage.xaml.cs.
Updated MainWindow to navigate to the new MainPage on load.
