+++ EduFlow.Desktop/Pages/MainForPage/MainPage.xaml
@@ -1,0 +1,70 @@
+﻿<Page x:Class="EduFlow.Desktop.Pages.MainForPage.MainPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:EduFlow.Desktop.Pages.MainForPage"
+      mc:Ignorable="d"
+      Title="MainPage">
+
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="80"/>
+            <RowDefinition Height="*"/>
+        </Grid.RowDefinitions>
+
+        <Grid Grid.Row="0"
+              Margin="10 5">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="*"/>
+                <ColumnDefinition Width="*"/>
+                <ColumnDefinition Width="*"/>
+                <ColumnDefinition Width="*"/>
+            </Grid.ColumnDefinitions>
+
+            <Border Grid.Column="0"
+                    CornerRadius="5"
+                    BorderThickness="0"
+                    Width="160"
+                    Background="Tomato">
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock x:Name="YourCoursesCount"
+                               Text="3"
+                               Foreground="White"
+                               FontSize="22"
+                               HorizontalAlignment="Center"
+                               FontWeight="Heavy"/>
+
+                    <TextBlock Text="Kurslar soni"
+                               Foreground="White"
+                               FontSize="14"
+                               HorizontalAlignment="Center"
+                               FontWeight="SemiBold"
+                               FontFamily="Jetbrains Mono"/>
+                </StackPanel>
+            </Border>
+
+            <Border Grid.Column="1"
+                    CornerRadius="5"
+                    BorderThickness="0"
+                    Width="160"
+                    Background="DeepSkyBlue">
+                <StackPanel VerticalAlignment="Center">
+                    <TextBlock x:Name="YourStudentsCount"
+                               Text="3"
+                               Foreground="White"
+                               FontSize="22"
+                               HorizontalAlignment="Center"
+                               FontWeight="Heavy"/>
+
+                    <TextBlock Text="O'quvchilar soni"
+                               Foreground="White"
+                               FontSize="14"
+                               HorizontalAlignment="Center"
+                               FontWeight="SemiBold"
+                               FontFamily="Jetbrains Mono"/>
+                </StackPanel>
+            </Border>
+        </Grid>
+    </Grid>
+</Page>
