--- EduFlow.Desktop/Pages/MainPage.xaml
@@ -1,13 +1,0 @@
-﻿<Page x:Class="EduFlow.Desktop.Pages.MainPage"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-      xmlns:local="clr-namespace:EduFlow.Desktop.Pages"
-      mc:Ignorable="d"
-      Title="MainPage">
-
-    <Grid>
-        
-    </Grid>
-</Page>
