+++ EduFlow.Desktop/Windows/MainWindow.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System.Windows;
+﻿using EduFlow.Desktop.Pages.MainForPage;
+using System.Windows;
 using System.Windows.Controls;
 
 namespace EduFlow.Desktop.Windows;
@@ -37,6 +38,8 @@ public partial class MainWindow : Window
     private void Window_Loaded(object sender, RoutedEventArgs e)
     {
         NormalButton.IsChecked = true;
+        MainPage page = new MainPage();
+        Navigate.Content = page;
     }
 
     private void CloseBtn_Click(object sender, RoutedEventArgs e)
